requirements.txt
psycopg2==2.4.5
us==0.9.0
django-phonenumber-field==0.7.2
jsonfield==1.0.3
geopy==1.9.1
django-statsd-mozilla==0.3.15
django-generic-m2m==0.3.0
Django==1.8.4
django-debug-toolbar==1.3.0
celery==3.1.18
flower==0.8.1
requests==2.7.0
git+https://github.com/django-tastypie/django-tastypie#egg=TastyPie
boto==2.38.0
ipdb==0.8.1
Pillow==2.9.0
django-storages==1.1.8
django-compressor==1.5
django-widget-tweaks==1.4.1
numpy==1.9.2
-e git+https://github.com/FireCARES/fire-risk.git#egg=fire-risk
django-queryset-csv==0.3.1
django-registration-redux==1.2
django-autocomplete-light==2.2.5
python-social-auth==0.2.14
django-multi-email-field==0.3.1
